% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cn_iecsc.R
\docType{data}
\name{cn_iecsc}
\alias{cn_iecsc}
\title{The China Inventory of Existing Chemical Substances and Chemicals}
\format{
A data frame with 37,069 rows and 1 variable:
\describe{
\item{cas_number}{(Character) Chemical Abstracts Service (CAS) registry
number}
}
}
\usage{
cn_iecsc
}
\description{
Note that this data was extracted from messy PDFs and is not
complete. In particular, (English) chemical names still need to be
extracted. For further information on the data sets, please visit the
official website from which the raw data originated:
\itemize{
\item https://www.mee.gov.cn/gkml/hbb/bgg/201301/t20130131_245810.htm
}
}
\keyword{datasets}
