% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ici.R
\docType{data}
\name{ici}
\alias{ici}
\title{The Australia Industrial Chemicals Inventory Data Set}
\format{
A data frame with 39,460 rows and 6 variables:
\describe{
\item{cr_no}{(Character) Chemical Abstracts Service (CAS) registry number}
\item{cas_no}{(Character) the chemical name}
\item{cas_name}{(Character) the associated hazard category}
\item{aicis_approved_chemical_name_aacn}{(Character) any associated signal words and pictogram codes}
\item{molecular_formula}{(Character) the harmonized hazard statement codes}
\item{version}{(Character) the harmonized hazard statements}
}
}
\source{
\url{https://www.industrialchemicals.gov.au/search-inventory}
}
\usage{
ici
}
\description{
The Australian Inventory of Industrial Chemicals (Inventory) is
a searchable database of around 40,000 chemicals that are being
manufactured or imported (introduced) into Australia for industrial use.

This downloadable version of the Inventory is provided for convenience
only and is not the official version of the Inventory. AICIS regularly
updates and makes changes to the Inventory that are not incorporated in
this downloadable version. The downloadable Inventory:
\itemize{
\item is a copy of the publicly available information on the Inventory as at
10 February 2022 and may not be complete, accurate and up-to-date
\item should not be relied on for the latest information about a chemical –
you must search the online Inventory to see if your chemical is listed
and what any terms of its listing are
\item does not include chemicals that cannot be disclosed to the public
because the terms are confidentially listed on the Inventory
\item does not include URL links to assessment and evaluation statements
}
}
\keyword{datasets}
