% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-my_cims.R
\docType{data}
\name{my_cims}
\alias{my_cims}
\title{The Malaysia Chemical Information Management System Inventory}
\format{
A data frame with 5,074 rows and 3 variables:
\describe{
\item{cas_no}{(Character) The Chemical Abstract Service (CAS) Registry
Number}
\item{chemical_name}{(Character) The substance name}
\item{iupac_name}{(Character) The IUPAC-conform substance name}
}
}
\usage{
my_cims
}
\description{
Note that the .tsv file needs to be manually downloaded
from the official website:
https://cims.dosh.gov.my/
}
\keyword{datasets}
