% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ncis.R
\docType{data}
\name{ncis}
\alias{ncis}
\title{The South Korea National Chemicals Information System Data Set}
\format{
A data frame with 46,111 rows and 4 variables:
\describe{
\item{cas_no}{(Character) Chemical Abstracts Service (CAS) registry number}
\item{chemical_name}{(Character) the chemical name}
\item{ke_no}{(Character) information on the approval status}
\item{date}{(Date) information on restrictions and exclusions}
}
}
\source{
\url{https://ncis.nier.go.kr/en/mttrList.do}
}
\usage{
ncis
}
\description{
When there is no matching result, or the substance is only
marked on the "OECD HPV chemicals", it is considered "Non-phase-in
substance" ("OECD HPV chemicals" means chemical substances that are
globally manufactured or imported over 1,000 tons.)

Please note that the information on salts of toxic substance or heavy
metal compound may not be found at NCIS. In this case, it is recommended
to consider relevant Public Notices of the Ministry of Environment for
understanding accurate regulatory information.

When the substance you search is a "hydrate" and other substance except
the hydrate is listed as phase-in substance, the substance you search is
not deemed as non-phase-in substance.

"Hydrate" of phase-in substance subject to registration is also deemed
as phase-in substance subject to registration.

Percentage in product and regulatory information of toxic substance,
substance subject to authorization, restricted substance or prohibited
substance are also available by clicking "Percentage & Regulatory
Information".

When "Substance Requiring Preparation for Accidents" under NIER's No. is
marked, it is recommended to find more information at "Korea Information
System for Chemical Management (http://kischem.nier.go.kr)."
}
\keyword{datasets}
