% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ph_emb_mc_07.R
\docType{data}
\name{ph_emb_mc_07}
\alias{ph_emb_mc_07}
\title{The Philippine Inventory of Chemicals and Chemical Substances EMB-MC
2020-007}
\format{
A data frame with 31 rows and 3 variables:
\describe{
\item{no}{(Integer) The list number of each substance}
\item{cas_registry_no}{(Character) Chemical Abstracts Service (CAS)
registry number}
\item{notified_chemical_name}{(Character) The associated chemical name}
}
}
\usage{
ph_emb_mc_07
}
\description{
Note that this data set was manually extracted from a messy
PDF. The entire procedure is documented in the "cleanventory" R package.
For further information on the data sets, please visit the official
website from which the raw data originated:
\itemize{
\item https://chemical.emb.gov.ph/?page_id=138
}
}
\keyword{datasets}
