% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-us_cdr_2016_napv.R
\docType{data}
\name{us_cdr_2016_napv}
\alias{us_cdr_2016_napv}
\title{The United States Chemical Data Reporting Inventory, 2016 National
Aggregate Production Volume}
\format{
A data frame with 8,768 rows and 2 variables:
\describe{
\item{chemical_id_number}{(Character) The Chemical Abstract Service (CAS)
Registry Number of the substance}
\item{stripped_chemical_id_number}{(Integer) The CAS Registry Number
without hyphens}
}
}
\usage{
us_cdr_2016_napv
}
\description{
Note that the .csv file needs to be manually downloaded
from the official website:
https://www.epa.gov/chemical-data-reporting
}
\keyword{datasets}
