% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-xn_spin.R
\docType{data}
\name{xn_spin}
\alias{xn_spin}
\title{The Substances in Preparations in Nordic Countries Inventory}
\format{
A data frame with 31,255 rows and 2 variables:
\describe{
\item{cas_no}{(Character) The Chemical Abstract Service (CAS)
Registry Number of the substance}
\item{substances}{(Character) The substance name}
}
}
\usage{
xn_spin
}
\description{
Note that the .tsv file needs to be manually downloaded
from the official website:
http://www.spin2000.net/spinmyphp/
}
\keyword{datasets}
